Fix NPE bug in JForumSecureSSO
When JForumSecureSSO asks for its cookie from JForum, the latter may
legally return null. JForumSecureSSO didn't check for this though
and as a result jforum would crash with NPE from time to time in
openthesaurus.gr.

This patch fixes the bug by checking and if ssoCookie is null, returns
false, invalidating the session.